requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
iniconfig==2.0.0
oauthlib==3.2.2
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.8.0
-e git+https://github.com/vkqapro/pythonRequestsAPI.git@901d8710d97141f74c6bbb05b7e2de685f4f1cf6#egg=Trello_API_Regression
urllib3==2.3.0
